requirements.txt
absl-py==0.10.0
aio-pika==6.7.1
aiofiles==0.6.0
aiohttp==3.6.3
aiormq==3.3.1
altair==4.1.0
APScheduler==3.6.3
argon2-cffi==20.1.0
astor==0.8.1
astunparse==1.6.3
async-generator==1.10
async-timeout==3.0.1
attr==0.3.1
attrs==20.2.0
backcall==0.2.0
base58==2.0.1
bleach==3.2.1
blinker==1.4
blis==0.7.3
boto3==1.16.25
botocore==1.19.25
cached-property==1.5.2
cachetools==4.1.1
catalogue==1.0.0
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
cloudpickle==1.4.1
colorama==0.4.4
colorclass==2.2.0
coloredlogs==14.0
colorhash==1.0.2
coremltools==4.0
cryptography==3.2.1
cycler==0.10.0
cymem==2.0.4
dataclasses==0.6
decorator==4.4.2
defusedxml==0.6.0
dm-tree==0.1.5
dnspython==1.16.0
docopt==0.6.2
easyocr==1.2.1
entrypoints==0.3
enum-compat==0.0.3
fbmessenger==6.0.0
filelock==3.0.12
future==0.18.2
gast==0.3.3
gitdb==4.0.5
GitPython==3.1.11
google-auth==1.23.0
google-auth-oauthlib==0.4.2
google-pasta==0.2.0
grpcio==1.34.0
h11==0.9.0
h5py==2.10.0
httpcore==0.11.1
httplib2==0.18.1
httptools==0.1.1
httpx==0.15.4
humanfriendly==9.0
idna==2.10
imageio==2.9.0
importlib-metadata==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.2
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.15.1
jsonpickle==1.4.2
jsonschema==3.2.0
jupyter-client==6.1.7
jupyter-core==4.7.0
jupyterlab-pygments==0.1.2
kafka-python==2.0.2
Keras==2.4.3
keras-contrib @ git+https://www.github.com/keras-team/keras-contrib.git@3fc5ef709e061416f4bc8a92ca3750c824b5d2b0
Keras-Preprocessing==1.1.2
kiwisolver==1.3.1
kraken==2.0.8
lxml==4.6.1
Markdown==3.3.3
MarkupSafe==1.1.1
matplotlib==3.3.3
mattermostwrapper==2.2
mistune==0.8.4
mpmath==1.1.0
multidict==4.7.6
murmurhash==1.0.4
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.1.5
numpy==1.18.5
oauth2client==4.1.3
oauthlib==3.1.0
opencv-python==4.4.0.46
opt-einsum==3.3.0
packaging==20.4
pamqp==2.3.0
pandas==1.1.4
pandocfilters==1.4.3
parso==0.7.1
pathtools==0.1.2
pdf2image==1.14.0
pickleshare==0.7.5
Pillow==8.0.1
plac==1.1.3
preshed==3.0.4
prometheus-client==0.9.0
prompt-toolkit==2.0.10
protobuf==3.14.0
psycopg2-binary==2.8.6
pyarrow==2.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pydeck==0.5.0
pydot==1.4.1
Pygments==2.7.2
PyJWT==1.7.1
pykwalify==1.7.0
pymongo==3.10.1
pyparsing==2.4.7
pyreadline==2.1
pyrsistent==0.17.3
pyspellchecker==0.5.5
pytesseract==0.3.6
python-bidi==0.4.2
python-crfsuite==0.9.7
python-dateutil==2.8.1
python-engineio==3.13.2
python-socketio==4.6.1
python-telegram-bot==12.8
pytz==2020.4
PyWavelets==1.1.1
pywin32==300
pywinpty==0.5.7
PyYAML==5.3.1
pyzmq==20.0.0
questionary==1.5.2
rasa==2.1.2
rasa-sdk==2.1.2
redis==3.5.3
regex==2020.9.27
requests==2.25.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
rocketchat-API==1.9.1
rsa==4.6
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
s3transfer==0.3.3
sacremoses==0.0.43
sanic==20.9.1
Sanic-Cors==0.10.0.post3
sanic-jwt==1.4.1
Sanic-Plugins-Framework==0.9.4.post1
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.4
Send2Trash==1.5.0
sentencepiece==0.1.91
sentry-sdk==0.19.4
seqeval==1.2.2
six==1.15.0
sklearn==0.0
sklearn-crfsuite==0.3.6
slackclient==2.9.3
smmap==3.0.4
sniffio==1.2.0
spacy==2.3.3
SQLAlchemy==1.3.20
srsly==1.0.4
streamlit==0.71.0
sympy==1.6.2
tabulate==0.8.7
tensorboard==2.4.0
tensorboard-plugin-wit==1.7.0
tensorflow==2.3.1
tensorflow-addons==0.11.2
tensorflow-estimator==2.3.0
tensorflow-hub==0.9.0
tensorflow-probability==0.11.1
termcolor==1.1.0
terminado==0.9.1
terminaltables==3.1.0
testpath==0.4.4
thinc==7.4.3
threadpoolctl==2.1.0
tifffile==2020.11.18
tokenizers==0.9.3
toml==0.10.2
toolz==0.11.1
torch==1.7.0+cpu
torchaudio==0.7.0
torchvision==0.8.1+cpu
tornado==6.1
tqdm==4.50.2
traitlets==5.0.5
transformers==3.5.1
twilio==6.45.4
typeguard==2.10.0
typing-extensions==3.7.4.3
tzlocal==2.1
ujson==3.2.0
urllib3==1.26.2
validators==0.18.1
wasabi==0.8.0
watchdog==0.10.4
wcwidth==0.2.5
webencodings==0.5.1
webexteamssdk==1.6
websockets==8.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wincertstore==0.2
wrapt==1.12.1
yarl==1.5.1
zipp==3.4.0
